Gently modifing how my messages are serialized cuz I'm from the streets.

+++ ROS#/Messages/std_msgs/Int16.cs
@@ -15,6 +15,6 @@ namespace Messages.std_msgs
 		[StructLayout(LayoutKind.Sequential, Pack = 1)]
 		public struct Int16
 		{
-			public int data;
+			public short data;
 		}
 }
